# Speedup x 2.15: Generate `__version__` at build to avoid slow `importlib.metadata` import