--- a/src/linkotron/__init__.py
+++ b/src/linkotron/__init__.py
@@ -4,11 +4,10 @@ CLI to format links
 
 from __future__ import annotations
 
-import importlib.metadata
 import re
 from typing import Any
 
-__version__ = importlib.metadata.version(__name__)
+from ._version import __version__ as __version__
 
 
 # https://github.com/nedbat/adventofcode2022/blob/main/day07.py