--- a/tox.ini
+++ b/tox.ini
@@ -28,3 +28,15 @@ pass_env =
     PRE_COMMIT_COLOR
 commands =
     pre-commit run --all-files --show-diff-on-failure
+
+[testenv:mypy]
+deps =
+    httpx
+    mypy==1.10.1
+    platformdirs
+    pyperclip
+    pytest
+    python-slugify
+    types-freezegun
+commands =
+    mypy . {posargs}