Like https://github.com/hugovk/tinytext/pull/195.

With Python 3.13.0rc2:

```sh
python -X importtime -c "import linktron" 2> import.log && tuna import.log
```

# `main`

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/369f40c0-9cf4-47fb-b3f0-56d30fa2fa6d">

# PR

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/3f3ec3fe-6051-487b-a2dc-2324a2cc2c7d">

# hyperfine

```console
❯ hyperfine --warmup 32 \
--prepare "git checkout rm-importlib.metadata" 'python3 -c "import linkotron # PR"' \
--prepare "git checkout main"                  'python3 -c "import linkotron # main"'
Benchmark 1: python3 -c "import linkotron # PR"
  Time (mean ± σ):      18.6 ms ±   1.3 ms    [User: 14.8 ms, System: 3.4 ms]
  Range (min … max):    17.7 ms …  25.6 ms    106 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs.

Benchmark 2: python3 -c "import linkotron # main"
  Time (mean ± σ):      40.0 ms ±   1.3 ms    [User: 32.5 ms, System: 6.9 ms]
  Range (min … max):    38.9 ms …  47.3 ms    60 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs.

Summary
  python3 -c "import linkotron # PR" ran
    2.15 ± 0.16 times faster than python3 -c "import linkotron # main"
```

---

Also move mypy from pre-commit to tox, so it runs when the version file has been generated.